backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0

# LangChain for RAG Pipeline
langchain==0.0.350
langchain-openai==0.0.2
langchain-community==0.0.10
langchain-pinecone==0.0.3

# OpenAI Integration
openai==1.3.7

# Vector Database
pinecone-client==2.2.4

# Data Processing and ML
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
sentence-transformers==2.2.2

# Text Processing
tiktoken==0.5.2
pypdf==3.17.1
python-docx==1.1.0
beautifulsoup4==4.12.2

# Environment and Configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Logging and Monitoring
structlog==23.2.0

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# CORS and Security
# CORS is handled by FastAPI built-in CORSMiddleware

# Development Tools
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
